requirements.txt
pandas
numpy
pytest
pyyaml
python-dotenv
